Introducing ConfigLib
Scripts now can open a config file, passing in a lua value (ie table). That value will be used as the defaults, and if a type mismatch occurs between the default and the file while (deep) parsing, the default value will be used. If a table has an unnamed value in a table, the entire table is assumed to be a json array rather than an object.

Partially addresses #26